chore(viewer): Rename folder resource to representation
