+++ projects/knora/viewer/src/lib/representation/index.ts
