+++ projects/knora/viewer/src/lib/representation/moving-image/moving-image.component.spec.ts
