+++ projects/knora/viewer/src/lib/representation/still-image/still-image.component.scss
